+++ duckscript_sdk/src/sdk/std/env/mod.rs
@@ -1,4 +1,5 @@
-mod get;
+mod env_to_map;
+mod get_env;
 mod get_home_dir;
 mod get_user_name;
 mod is_windows;
@@ -7,8 +8,8 @@ mod os_name;
 mod os_release;
 mod os_version;
 mod print_current_directory;
-mod set;
 mod set_current_directory;
+mod set_env;
 mod uname;
 mod unset;
 
@@ -21,7 +22,8 @@ static PACKAGE: &str = "env";
 pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptError> {
     let package = pckg::concat(parent, PACKAGE);
 
-    commands.set(get::create(&package))?;
+    commands.set(env_to_map::create(&package))?;
+    commands.set(get_env::create(&package))?;
     commands.set(get_home_dir::create(&package))?;
     commands.set(get_user_name::create(&package))?;
     commands.set(is_windows::create(&package)?)?;
@@ -30,8 +32,8 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(os_release::create(&package))?;
     commands.set(os_version::create(&package))?;
     commands.set(print_current_directory::create(&package))?;
-    commands.set(set::create(&package))?;
     commands.set(set_current_directory::create(&package))?;
+    commands.set(set_env::create(&package))?;
     commands.set(uname::create(&package)?)?;
     commands.set(unset::create(&package))?;